ARIN5202-NLP-FinalProject-main/requirements.txt
# Core Web Framework
Flask==3.1.2
python-dotenv==1.2.1

# Finance Data
yfinance>=0.2.66

# Document Processing
PyPDF2==3.0.1
python-docx==1.2.0
pillow==11.3.0
pytesseract==0.3.13

# NLP & Retrieval
spacy==3.8.9
rank-bm25==0.2.2
sentence-transformers==5.1.2
faiss-cpu==1.13.0

# Vision AI via Ollama
ollama==0.4.8

# Data Processing
datasets==4.4.1
requests==2.32.5
